BufferQueue abandoned bug + modularization of code
