+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
 
         </activity>
 
-        <activity android:name=".ui.Camera" />
+        <activity android:name=".ui.Camera.CameraView" />
         <activity
             android:name=".ui.sharing.PlaylistActivity"
             android:exported="false"
